+++ src/main/java/ga/ganma/minigames/missions/Mission4Time.java
@@ -1,0 +1,38 @@
+package ga.ganma.minigames.missions;
+
+import ga.ganma.minigames.TosoNow;
+import ga.ganma.minigames.commands.TosoCommand;
+import ga.ganma.minigames.Mission;
+import ga.ganma.minigames.listeners.GameListeners;
+
+import org.bukkit.Bukkit;
+import org.bukkit.ChatColor;
+import org.bukkit.Material;
+import org.bukkit.Sound;
+import org.bukkit.entity.Player;
+import org.bukkit.scheduler.BukkitRunnable;
+
+public class Mission4Time extends BukkitRunnable {
+
+	@Override
+	public void run() {
+		if (TosoNow.gameTime == Mission.mission4t) {
+			Mission.isMission = false;
+			this.cancel();
+			GameListeners.chat = true;
+			for (Player allp : Bukkit.getServer().getOnlinePlayers()) {
+				allp.sendTitle(ChatColor.RED + "誰も押さなかったため、賞金単価は上がらなかった...", "", 0, 100, 0);
+				allp.playSound(allp.getLocation(), Sound.UI_BUTTON_CLICK, 1, 2);
+			}
+			TosoCommand.world.getBlockAt(Mission.blockL).setType(Material.AIR);
+		} else if (!Mission.isMission) {
+			this.cancel();
+			GameListeners.chat = true;
+			for (Player allp : Bukkit.getServer().getOnlinePlayers()) {
+				allp.sendTitle(ChatColor.RED + "ハンター強化！", Mission.CLEARp.getName() + "により、単価が300円に上がったがハンターが強化された！", 0,
+						100, 0);
+				allp.playSound(allp.getLocation(), Sound.UI_BUTTON_CLICK, 1, 2);
+			}
+		}
+	}
+}
